+++ wp-content/uploads/civicrm/ext/mjwshared/settings/mjwshared.setting.php
@@ -16,7 +16,7 @@ return [
     'name' => 'mjwshared_refundpaymentui',
     'type' => 'Boolean',
     'html_type' => 'checkbox',
-    'default' => 1,
+    'default' => TRUE,
     'is_domain' => 1,
     'is_contact' => 0,
     'title' => E::ts('Enable refund payment via UI?'),
@@ -24,8 +24,24 @@ return [
     For more detail see the <a href="%1">Refund documentation</a>', [1 => 'https://docs.civicrm.org/mjwshared/en/latest/refunds/']),
     'html_attributes' => [],
     'settings_pages' => [
-      'stripe' => [
+      'mjwshared' => [
         'weight' => 21,
+      ]
+    ],
+  ],
+  'mjwshared_disablerecordrefund' => [
+    'name' => 'mjwshared_disablerecordrefund',
+    'type' => 'Boolean',
+    'html_type' => 'checkbox',
+    'default' => FALSE,
+    'is_domain' => 1,
+    'is_contact' => 0,
+    'title' => E::ts('Disable the "Record Refund" link on edit contribution'),
+    'description' => E::ts('By default CiviCRM includes a "Record Refund" link on edit contribution. This can be confusing when our payment refund UI is enabled because the contribution "Record Refund" does not communicate with the payment processor.'),
+    'html_attributes' => [],
+    'settings_pages' => [
+      'mjwshared' => [
+        'weight' => 22,
       ]
     ],
   ],
@@ -40,7 +56,7 @@ return [
     'description' => E::ts('Enables debug logging to browser console for javascript based payment processors.'),
     'html_attributes' => [],
     'settings_pages' => [
-      'stripe' => [
+      'mjwshared' => [
         'weight' => 99,
       ]
     ],